requirements.txt
numpy==1.17.2
torchvision==0.5.0
torch==1.4.0
opencv_python==4.2.0.34
scipy==1.3.1
matplotlib==3.1.1
tqdm==4.36.1
Pillow==7.1.2
scikit_learn==0.23.1
